perf(app): split cm chunks
