+++ apps/codeimage/src/components/CustomEditor/CustomEditor.tsx
@@ -2,7 +2,30 @@ import {SUPPORTED_LANGUAGES} from '@codeimage/config';
 import {getActiveEditorStore} from '@codeimage/store/editor/createActiveEditor';
 import {getRootEditorStore} from '@codeimage/store/editor/createEditors';
 import {getThemeStore} from '@codeimage/store/theme/theme.store';
-import {EditorView, lineNumbers} from '@codemirror/view';
+import {
+  autocompletion,
+  closeBrackets,
+  closeBracketsKeymap,
+  completionKeymap,
+} from '@codemirror/autocomplete';
+import {
+  defaultKeymap,
+  history,
+  historyKeymap,
+  indentWithTab,
+} from '@codemirror/commands';
+import {bracketMatching, indentOnInput} from '@codemirror/language';
+import {EditorState, Extension} from '@codemirror/state';
+import {
+  crosshairCursor,
+  drawSelection,
+  dropCursor,
+  EditorView,
+  highlightSpecialChars,
+  keymap,
+  lineNumbers,
+  rectangularSelection,
+} from '@codemirror/view';
 import {SUPPORTED_FONTS} from '@core/configuration/font';
 import {ReplaySubject} from 'rxjs';
 import {createCodeMirror} from 'solid-codemirror';
@@ -13,19 +36,39 @@ import {
   createResource,
   onCleanup,
 } from 'solid-js';
-import {createCustomFontExtension} from './custom-font-extension';
-import {fixCodeMirrorAriaRole} from './fix-cm-aria-roles-lighthouse';
 import {observeFocusExtension} from './observe-focus-extension';
+
+interface CustomFontExtensionOptions {
+  fontName: string;
+  fontWeight: number;
+}
+
+const EDITOR_BASE_SETUP: Extension = [
+  highlightSpecialChars(),
+  drawSelection(),
+  dropCursor(),
+  EditorState.allowMultipleSelections.of(true),
+  indentOnInput(),
+  bracketMatching(),
+  closeBrackets(),
+  autocompletion(),
+  rectangularSelection(),
+  crosshairCursor(),
+  history(),
+  keymap.of([
+    ...closeBracketsKeymap,
+    ...defaultKeymap,
+    ...completionKeymap,
+    ...historyKeymap,
+    indentWithTab,
+  ]),
+];
 
 export default function CustomEditor() {
   let editorEl!: HTMLDivElement;
   const {themeArray: themes} = getThemeStore();
 
-  const [basicSetup] = createResource(() =>
-    import('./basic-setup').then(e => e.EDITOR_BASE_SETUP),
-  );
-
-  fixCodeMirrorAriaRole(() => editorEl);
+  // fixCodeMirrorAriaRole(() => editorEl);
   const destroy$ = new ReplaySubject<void>(1);
   const languages = SUPPORTED_LANGUAGES;
   const fonts = SUPPORTED_FONTS;
@@ -92,6 +135,23 @@ export default function CustomEditor() {
     },
   });
 
+  const createCustomFontExtension = (
+    options: CustomFontExtensionOptions,
+  ): Extension => {
+    return EditorView.theme({
+      '.cm-content *': {
+        fontFamily: `${options.fontName}, monospace`,
+        fontWeight: options.fontWeight,
+        fontVariantLigatures: 'normal',
+      },
+      '.cm-gutters': {
+        fontFamily: `${options.fontName}, monospace`,
+        fontWeight: 400,
+        fontVariantLigatures: 'normal',
+      },
+    });
+  };
+
   const customFontExtension = () =>
     createCustomFontExtension({
       fontName:
@@ -122,9 +182,6 @@ export default function CustomEditor() {
 
   createEffect(() => {
     batch(() =>
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      // TODO: to fix type deep instantion
       setOptions({
         extensions: [
           baseTheme,
@@ -134,7 +191,7 @@ export default function CustomEditor() {
           currentLanguage() || [],
           currentTheme(),
           editorOptions.showLineNumbers ? lineNumbers() : [],
-          basicSetup() || [],
+          EDITOR_BASE_SETUP,
         ],
       }),
     );
