--- apps/codeimage/src/components/CustomEditor/basic-setup.ts
@@ -1,43 +1,0 @@
-import {
-  autocompletion,
-  closeBrackets,
-  closeBracketsKeymap,
-  completionKeymap,
-} from '@codemirror/autocomplete';
-import {
-  defaultKeymap,
-  history,
-  historyKeymap,
-  indentWithTab,
-} from '@codemirror/commands';
-import {bracketMatching, indentOnInput} from '@codemirror/language';
-import {EditorState, Extension} from '@codemirror/state';
-import {
-  crosshairCursor,
-  drawSelection,
-  dropCursor,
-  highlightSpecialChars,
-  keymap,
-  rectangularSelection,
-} from '@codemirror/view';
-
-export const EDITOR_BASE_SETUP: Extension = [
-  highlightSpecialChars(),
-  drawSelection(),
-  dropCursor(),
-  EditorState.allowMultipleSelections.of(true),
-  indentOnInput(),
-  bracketMatching(),
-  closeBrackets(),
-  autocompletion(),
-  rectangularSelection(),
-  crosshairCursor(),
-  history(),
-  keymap.of([
-    ...closeBracketsKeymap,
-    ...defaultKeymap,
-    ...completionKeymap,
-    ...historyKeymap,
-    indentWithTab,
-  ]),
-];
