+++ apps/codeimage/src/components/CustomEditor/supportsLineWrap.ts
@@ -1,0 +1,3 @@
+import {EditorView} from '@codemirror/view';
+
+export const supportsLineWrap = EditorView.lineWrapping;
